Add dynamic-login element
This element allows one to send a kernel parameter (via PXE or iPXE)
"rootpwd" or "sshkey", which will install a root password or SSH key
during boot. In our case, we can use rootpwd to set up a local root user
so we can log in during a clean step.